[#3] refactor: utils folder renamed to services
